+++ product_harmonized_system/models/product_template.py
@@ -2,7 +2,7 @@
 ##############################################################################
 #
 #    Copyright (C) 2011-2015 Akretion (http://www.akretion.com)
-#    Copyright (C) 2015 Noviat (http://www.noviat.com)
+#    Copyright (C) 2009-2015 Noviat (http://www.noviat.com)
 #    @author Alexis de Lattre <alexis.delattre@akretion.com>
 #    @author Luc de Meyer <info@noviat.com>
 #
@@ -46,28 +46,6 @@ class ProductTemplate(models.Model):
             res = self.hs_code_id
         elif self.categ_id:
             res = self.categ_id.get_hs_code_recursively()
-        else:
-            res = None
-        return res
-
-
-class ProductCategory(models.Model):
-    _inherit = "product.category"
-
-    hs_code_id = fields.Many2one(
-        'hs.code', string='H.S. Code',
-        company_dependent=True, ondelete='restrict',
-        help="Harmonised System Code. If this code is not "
-        "set on the product itself, it will be read here, on the "
-        "related product category.")
-
-    @api.multi
-    def get_hs_code_recursively(self):
-        self.ensure_one()
-        if self.hs_code_id:
-            res = self.hs_code_id
-        elif self.parent_id:
-            res = self.parent_id.get_hs_code_recursively()
         else:
             res = None
         return res
