+++ product_harmonized_system/views/product_template.xml
@@ -22,19 +22,5 @@
       </field>
     </record>
 
-    <!-- Product category form view -->
-    <record id="product_category_form_view" model="ir.ui.view">
-      <field name="name">hs_code.product.category.form</field>
-      <field name="model">product.category</field>
-      <field name="inherit_id" ref="product.product_category_form_view"/>
-      <field name="arch" type="xml">
-        <xpath expr="//group[@name='parent']" position="inside">
-          <group name="hs_code" string="Import/Export Properties" colspan="2">
-            <field name="hs_code_id"/>
-          </group>
-        </xpath>
-      </field>
-    </record>
-
   </data>
 </openerp>
